♻️ tiptap header 부분 컴포넌트화
